// Frontend/ClientApp/app/shared/models/user.ts
export class User {
    id: String;
    firstname: String;
    job: String;
    password:string
    plantId: Number;
    sectorId: Number;
    jobId: Number;
    jodId:Number;
    name:String;
    phoneNumber: String;
    plant: String;
    plantID: Number;
    sector:String;
    sectorID:Number;
    roles:[];
    surname: String;
    email: string;
    username:String;
    address: String;
    base64Profile: string;
}